pyproject.toml
[project]
name = "fifa-table"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pygithub>=2.5.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]
